sharjeel_riaz/PCapp/requirements.txt
aws-cdk.core==1.135.0
aws-cdk.assertions==1.135.0

